+++ types.ts
@@ -1472,6 +1472,8 @@ export interface SkillEntrypointConfig {
 
 /**
  * Entrypoints that skills can be invoked from.
+ *
+ * @deprecated Use {@link PackDefinitionBuilder.setChatSkill} instead.
  */
 export interface SkillEntrypoints {
   /**