+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This changelog keeps track of all changes to the Packs SDK. We follow convention
 
 ## [Unreleased]
 
+## [1.13.0] - 2026-02-06
+
 ### Added
 
 - Added `setBenchInitializationSkill`, which allows the agent to run a skill immediately when the agent icon is clicked on in the bench.
@@ -988,7 +990,7 @@ await myHelper(context);
 
 - Beginning of alpha versioning.
 
-[unreleased]: https://github.com/coda/packs-sdk/compare/v1.12.5...HEAD
+[unreleased]: https://github.com/coda/packs-sdk/compare/v1.13.0...HEAD
 [1.7.5]: https://github.com/coda/packs-sdk/compare/v1.7.1...v1.7.5
 [1.7.4]: https://github.com/coda/packs-sdk/compare/v1.7.1...v1.7.4
 [1.7.3]: https://github.com/coda/packs-sdk/compare/v1.7.1...v1.7.3
@@ -1055,6 +1057,7 @@ await myHelper(context);
 [1.11.1]: https://github.com/coda/packs-sdk/compare/v1.7.8...v1.11.1
 [1.11.0]: https://github.com/coda/packs-sdk/compare/v1.7.8...v1.11.0
 [1.12.0]: https://github.com/coda/packs-sdk/compare/v1.7.8...v1.12.0
+[1.13.0]: https://github.com/coda/packs-sdk/compare/v1.7.8...v1.13.0
 [1.12.5]: https://github.com/coda/packs-sdk/compare/v1.7.8...v1.12.5
 [1.12.4]: https://github.com/coda/packs-sdk/compare/v1.7.8...v1.12.4
 [1.12.3]: https://github.com/coda/packs-sdk/compare/v1.7.8...v1.12.3